修复新伤害计算流程

+++ src/main/java/me/onebone/actaeon/entity/MovingEntity.java
@@ -446,12 +446,12 @@ abstract public class MovingEntity extends EntityCreature implements IMovingEnti
 		super.setOnFire(seconds);
 	}
 
-	protected double calculateEnchantmentProtectionFactor(Item item, EntityDamageEvent source) {
+	protected int calculateEnchantmentProtectionFactor(Item item, EntityDamageEvent source) {
 		if (!item.hasEnchantments()) {
 			return 0;
 		}
 
-		double epf = 0;
+		int epf = 0;
 
 		for (Enchantment ench : item.getEnchantments()) {
 			epf += ench.getProtectionFactor(source);
